1、move some function with sharding value to ParsingSQLRouter class(将sharding value相关方法移到ParsingSQLRouter中) 2、rename package name io.shardingjdbc.core.routing.condition to io.shardingjdbc.core.routing.sharding(重命名包io.shardingjdbc.core.routing.condition为io.shardingjdbc.core.routing.sharding) 3、modify sharding value package class(修改sharding value的包装类)
