+++ sharding-core/src/main/java/io/shardingjdbc/core/routing/SQLRouteResult.java
@@ -18,7 +18,7 @@
 package io.shardingjdbc.core.routing;
 
 import io.shardingjdbc.core.parsing.parser.sql.SQLStatement;
-import io.shardingjdbc.core.routing.condition.GeneratedKey;
+import io.shardingjdbc.core.routing.sharding.GeneratedKey;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
