--- sharding-core/src/main/java/io/shardingjdbc/core/routing/condition/ShardingConditions.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 1999-2015 dangdang.com.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * </p>
- */
-
-package io.shardingjdbc.core.routing.condition;
-
-import io.shardingjdbc.core.api.algorithm.sharding.ShardingValues;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-import java.util.List;
-
-/**
- * Sharding conditions.
- *
- * @author zhangliang
- */
-@RequiredArgsConstructor
-@Getter
-public final class ShardingConditions {
-    
-    private final List<Object> parameters;
-    
-    private final ShardingValues shardingValues;
-    
-    private final GeneratedKey generatedKey;
-}
