+++ sharding-core/src/main/java/io/shardingjdbc/core/routing/type/standard/StandardRoutingEngine.java
@@ -21,12 +21,10 @@ import com.google.common.base.Objects;
 import com.google.common.base.Optional;
 import com.google.common.base.Preconditions;
 import io.shardingjdbc.core.api.algorithm.sharding.ShardingValue;
-import io.shardingjdbc.core.api.algorithm.sharding.ShardingValueUnit;
-import io.shardingjdbc.core.api.algorithm.sharding.ShardingValues;
 import io.shardingjdbc.core.hint.HintManagerHolder;
 import io.shardingjdbc.core.hint.ShardingKey;
-import io.shardingjdbc.core.routing.condition.GeneratedKey;
-import io.shardingjdbc.core.routing.condition.ShardingConditions;
+import io.shardingjdbc.core.routing.sharding.ShardingCondition;
+import io.shardingjdbc.core.routing.sharding.ShardingConditions;
 import io.shardingjdbc.core.routing.strategy.ShardingStrategy;
 import io.shardingjdbc.core.routing.type.RoutingEngine;
 import io.shardingjdbc.core.routing.type.RoutingResult;
@@ -55,10 +53,10 @@ public final class StandardRoutingEngine implements RoutingEngine {
     
     private final ShardingConditions shardingConditions;
     
-    public StandardRoutingEngine(final ShardingRule shardingRule, final List<Object> parameters, final String logicTableName, final ShardingValues shardingValues, final GeneratedKey generatedKey) {
+    public StandardRoutingEngine(final ShardingRule shardingRule, final List<Object> parameters, final String logicTableName, final ShardingConditions shardingConditions) {
         this.shardingRule = shardingRule;
         this.logicTableName = logicTableName;
-        shardingConditions = new ShardingConditions(parameters, shardingValues, generatedKey);
+        this.shardingConditions = shardingConditions;
     }
     
     @Override
@@ -77,13 +75,13 @@ public final class StandardRoutingEngine implements RoutingEngine {
     private List<ShardingValue> getDatabaseShardingValues(final TableRule tableRule) {
         ShardingStrategy strategy = shardingRule.getDatabaseShardingStrategy(tableRule);
         return HintManagerHolder.isUseShardingHint()
-                ? getDatabaseShardingValuesFromHint(strategy.getShardingColumns()) : findShardingValues(strategy.getShardingColumns());
+                ? getDatabaseShardingValuesFromHint(strategy.getShardingColumns()) : getShardingValues(strategy.getShardingColumns());
     }
     
     private List<ShardingValue> getTableShardingValues(final TableRule tableRule) {
         ShardingStrategy strategy = shardingRule.getTableShardingStrategy(tableRule);
         return HintManagerHolder.isUseShardingHint()
-                ? getTableShardingValuesFromHint(strategy.getShardingColumns()) : findShardingValues(strategy.getShardingColumns());
+                ? getTableShardingValuesFromHint(strategy.getShardingColumns()) : getShardingValues(strategy.getShardingColumns());
     }
     
     private List<ShardingValue> getDatabaseShardingValuesFromHint(final Collection<String> shardingColumns) {
@@ -108,12 +106,12 @@ public final class StandardRoutingEngine implements RoutingEngine {
         return result;
     }
     
-    private List<ShardingValue> findShardingValues(Collection<String> shardingColumns) {
+    private List<ShardingValue> getShardingValues(Collection<String> shardingColumns) {
         List<ShardingValue> result = new LinkedList<>();
-        if (!shardingConditions.getShardingValues().isEmpty()) {
-            Optional<ShardingValueUnit> shardingValueUnit = shardingConditions.getShardingValues().get(0);
-            if (shardingValueUnit.isPresent()) {
-                for (ShardingValue each : shardingValueUnit.get().getShardingValues()) {
+        if (!shardingConditions.isEmpty()) {
+            Optional<ShardingCondition> shardingCondition = shardingConditions.get(0);
+            if (shardingCondition.isPresent()) {
+                for (ShardingValue each : shardingCondition.get().getShardingValues()) {
                     if (Objects.equal(each.getLogicTableName(), logicTableName) && shardingColumns.contains(each.getColumnName())) {
                         result.add(each);
                     }
